requirements.txt
alembic==1.8.0
asttokens==2.0.5
attrs==21.4.0
Automat==20.2.0
backcall==0.2.0
bcrypt==3.2.2
blinker==1.4
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cloudinary==1.29.0
constantly==15.1.0
cryptography==37.0.2
cssselect==1.1.0
decorator==5.1.1
dnspython==2.2.1
email-validator==1.2.1
entrypoints==0.4
executing==0.8.3
filelock==3.7.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
incremental==21.3.0
ipykernel==4.8.2
ipython==8.4.0
itemadapter==0.6.0
itemloaders==1.0.4
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
jupyter-client==7.2.0
jupyter-core==4.10.0
lxml==4.9.0
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
parsel==1.6.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
Protego==0.2.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
Pygments==2.12.0
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pyzmq==17.0.0
queuelib==1.6.2
requests==2.28.0
requests-file==1.5.1
Scrapy==2.6.1
service-identity==21.1.0
six==1.16.0
SQLAlchemy==1.4.37
stack-data==0.3.0
stripe==3.4.0
tldextract==3.3.0
tornado==6.1
traitlets==5.3.0
Twisted==22.4.0
typing_extensions==4.2.0
urllib3==1.26.9
w3lib==1.22.0
wcwidth==0.2.5
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
zope.interface==5.4.0
